factorize components
